Hide Types.resolve.
